Added controls to only do backups when the database is modified.
The main purpose for this is to protect against a bug where a user saves very
quickly.  There may still be a race condition, but there is now a flag that
makes it so we only do backups when there is a reason to (database modified or
user explicitly requests it).