Fix exception when backup file already exists.
The default naming uses seconds, so if a user were to hit save more than once in
a second we would try and copy a file to a location that already exists.  This
causes an exception to the user.  The solution is to just skip the backup if the
file already exists.